test : move unit test to domain
