+++ src/test/java/nextstep/courses/infrastructure/CourseRepositoryTest.java
@@ -28,8 +28,6 @@ public class CourseRepositoryTest {
 
     private CourseRepository courseRepository;
 
-    private final DateTimeFormatter formatter = DateTimeFormatter.ofPattern("yyyy-MM-dd HH:mm:ss");
-
     @BeforeEach
     void setUp() {
         courseRepository = new JdbcCourseRepository(jdbcTemplate);
@@ -43,19 +41,5 @@ public class CourseRepositoryTest {
         Course savedCourse = courseRepository.findById(1L);
         assertThat(course.getTitle()).isEqualTo(savedCourse.getTitle());
         LOGGER.debug("Course: {}", savedCourse);
-    }
-
-    @Test
-    void term() {
-        Course course = new Course("TDD, 클린 코드 with Java", 1L);
-
-        List<LocalDateTime> expected = Arrays.asList(
-                LocalDateTime.parse("2022-01-01 11:11:11", formatter),
-                LocalDateTime.parse("2022-01-01 11:11:11", formatter)
-        );
-        course.putTerms(expected.get(0), expected.get(1));
-        List<LocalDateTime> actual = course.terms();
-
-        assertThat(expected).isEqualTo(actual);
     }
 }
